ci: ignore chromes outside ".gitlab/chrome" dir in terminate_chrome_processes
- this script now filters the processes by path so that user's instance of chrome.exe are unaffected.

+++ .gitlab/ci/terminate_chrome_processes.ps1
@@ -5,8 +5,11 @@ $chromeProcesses = Get-Process -Name "chrome" -ErrorAction SilentlyContinue
 if ($chromeProcesses) {
     # Iterate through each process and kill it
     foreach ($process in $chromeProcesses) {
-        Stop-Process -Id $process.Id -Force -ErrorAction SilentlyContinue
-        Write-Host "Killed process $($process.Name) with ID $($process.Id)"
+        # Only terminates chrome processes which started from CI.
+        if ("$($process.Path)" -Like "*\.gitlab\chrome\chrome.exe") {
+            Stop-Process -Id $process.Id -Force -ErrorAction SilentlyContinue
+            Write-Host "Killed process $($process.Name) with ID $($process.Id)"
+        }
     }
 } else {
     Write-Host "No chrome processes found."
